Changes to Algorithm Check --------------------------------------
-Added a dictionary component which incremented each penalty for each team. This allowed me to quickly compare my season long results to those on NHL.com. I found that 3 teams had one extra penalty. This lead to the Nashville, Ottawa, and Edmonton lists.

When looking over the results of the three lists, it was determined that the Edmonton list had an out of date file for the 12/31 game. The Ottawa game on 10/16 had a duplicate entry on the page for Mark Borowiecki. The Nashville game on 11/02/2014 had a duplicate entry for Shea Weber.

This to me confirms that my algorithm is able to every penalty that the NHL site lists.

In terms of time, this program was able to execute over 1230 HTML files and store the data in 38 seconds. This is the worst case scenario of processing the data.

TODO:
- Implement a way for the program to remember when it was last ran and only run through the games that have occurred between.
- Implement a way to store and retrieve all of the Penalty data. This might mean writing out all of the penalties to a file, then reading them in upon load.
- Implement a way to search the penalty list based on player, team, infraction, opponent, location and referees.

The last process change will be to have the penalty list run nightly and have the search capabilities happen upon user request.