Since the NHL redesigned their website using the MLB Advanced Stats group, I had to redesign a lot of the stuff to get the data I wanted out of it. Once this branch is merged with the trunk, it will effectively deprecate all the following files:
* scanning_games.py
* Teams.py
* gameDayProcessor.py

However, before I can reintegrate it, I need to change how it is currently formatting the date to be consistent with how the gameDayProcessor was doing it in the past. Additionally, I should probably change the home/away status. It doesn't make any sense to keep it as a boolean.